user verification working

+++ middlewares/users/userValidators.js
@@ -9,7 +9,7 @@ const addUserValidators = [
   check("name")
     .isLength({ min: 1 })
     .withMessage("Name is required")
-    .isAlpha("en-us", { ignore: " -" })
+    .isAlpha("en-US", { ignore: " -" })
     .withMessage("Name must not contain anything other than alphabet")
     .trim(),
   check("email")
@@ -54,7 +54,8 @@ const addUservalidationHandler = (req, res, next) => {
   if (Object.keys(mappedErrors).length === 0) {
     next();
   } else {
-    if (req.files.length > 0) {
+    // delete uploaded files
+    if (req.files && req.files.length > 0) {
       const { filename } = req.files[0];
       unlink(
         path.join(__dirname, `../../public/uploads/avatars/${filename}`),
@@ -63,7 +64,10 @@ const addUservalidationHandler = (req, res, next) => {
         },
       );
     }
-    res.status(500).json({
+
+    console.log("Validation Errors:", mappedErrors);
+
+    res.status(400).json({
       errors: mappedErrors,
     });
   }
